fix: error when trying to build
Here's the error I was getting
usePrismicPreview.ts(164,7): semantic error TS2352: Conversion of type 'Pick<PluginOptions, "repositoryName" | "accessToken" | "linkResolver" | "htmlSerializer" | "fetchLinks" | "lang" | "typePathsFilenamePrefix"> & { ...; } & { ...; }' to type 'BrowserPluginOptions' may be a mistake because neither type sufficiently overlaps with the other. If this was intentional, convert the expression to 'unknown' first.
  Property 'schemas' is missing in type 'Pick<PluginOptions, "repositoryName" | "accessToken" | "linkResolver" | "htmlSerializer" | "fetchLinks" | "lang" | "typePathsFilenamePrefix"> & { ...; } & { ...; }' but required in type 'Pick<PluginOptions, "repositoryName" | "accessToken" | "fetchLinks" | "lang" | "typePathsFilenamePrefix" | "releaseID" | "schemas">'.

This can also be fixed by adding schemas to the Options type in usePRismicPreview.ts